Crate chromascope

Chromascope is an easy-to-use GUI application designed to handle and read mzML mass spectrometry data.

The crate consists of three main modules:

gui.rs: This module contains the implementation of the graphical user interface (GUI) using the egui library.

parser.rs: This module handles the parsing and processing of the mzML data files.

plotting_parameters.rs: This module defines the parameters and settings for the data plotting functionality.

Modules

gui 🔒

MzViewerApp Module

parser 🔒

Parses and prepares MzML files for plotting

plotting_parameters 🔒

Line Properties Module

Functions

main 🔒